Change directory structure + package definition
