+++ CMakeLists.txt
@@ -28,7 +28,6 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
 
 set(CMAKE_MODULE_PATH ${PROJECT_SOURCE_DIR}/cmake)
-set(PONCA_ROOT_DIR ${PROJECT_SOURCE_DIR}/src)
 
 # guard against bad build-type strings (source: Eigen)
 
@@ -78,13 +77,11 @@ message(STATUS "Eigen3 root path: ${EIGEN3_INCLUDE_DIR}")
 ################################################################################
 # Generate install target                                                      #
 ################################################################################
-set(Ponca_INSTALL_INC_DIR "${CMAKE_INSTALL_PREFIX}/include")
-set(Ponca_INSTALL_LIB_DIR "${CMAKE_INSTALL_PREFIX}/lib")
 set(Ponca_EXPORT_TARGET_DIR \${CMAKE_CURRENT_LIST_DIR})
 
-install(EXPORT PoncaTargets
+install(EXPORT CoreTargets
   FILE
-    PoncaTargets.cmake
+    PoncaCoreTargets.cmake
   NAMESPACE
     Ponca::
   DESTINATION
@@ -134,7 +131,13 @@ install(
 ################################################################################
 # Add core code source, so it is visible in dev environements                  #
 ################################################################################
-add_subdirectory( src )
+
+set(PONCA_src_ROOT "${CMAKE_CURRENT_SOURCE_DIR}/")
+include(PoncaConfigureCore)
+
+install(DIRECTORY ${PONCA_src_ROOT}/Ponca
+    DESTINATION include/
+    PATTERN "*~" EXCLUDE)
 
 ################################################################################
 # API documentation with Doxygen                                               #
