+++ Ponca/Core
@@ -1,0 +1,33 @@
+/*
+ This Source Code Form is subject to the terms of the Mozilla Public
+ License, v. 2.0. If a copy of the MPL was not distributed with this
+ file, You can obtain one at http://mozilla.org/MPL/2.0/.
+*/
+
+#pragma once
+
+// Include Ponca Core components
+#include "src/Core/defines.h"
+#include "src/Core/enums.h"
+#include "src/Core/basket.h"
+
+#include "src/Core/weightKernel.h"
+#include "src/Core/weightFunc.h"
+
+#include "src/Core/plane.h"
+#include "src/Core/meanPlaneFit.h"
+#include "src/Core/covariancePlaneFit.h"
+#include "src/Core/mongePatch.h"
+
+#include "src/Core/sphereFit.h"
+#include "src/Core/orientedSphereFit.h"
+#include "src/Core/mlsSphereFitDer.h"
+#include "src/Core/curvature.h"
+#include "src/Core/gls.h"
+
+#include "src/Core/curvatureEstimation.h"
+
+// not supported on cuda
+#ifndef __CUDACC__
+# include "src/Core/unorientedSphereFit.h"
+#endif
