+++ Ponca/src/Core/algebraicSphere.hpp
