+++ Ponca/src/Core/covariancePlaneFit.hpp
