+++ Ponca/src/Core/curvature.hpp
