+++ Ponca/src/Core/curvatureEstimation.hpp
