+++ Ponca/src/Core/meanPlaneFit.hpp
