+++ Ponca/src/Core/mlsSphereFitDer.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include <Ponca/Core/defines.h>
+#include "./defines.h"
 
 namespace Ponca
 {
