+++ Ponca/src/Core/orientedSphereFit.hpp
