+++ Ponca/src/Core/sphereFit.h
@@ -6,7 +6,8 @@ file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 #pragma once
 
-#include <Ponca/Core/algebraicSphere.h>
+#include "./algebraicSphere.h"
+
 
 namespace Ponca
 {
