+++ Ponca/src/Core/sphereFit.hpp
