+++ Ponca/src/Core/unorientedSphereFit.hpp
