+++ Ponca/src/Core/weightFunc.hpp
