+++ examples/cpp/pcl/pcl_wrapper.h
@@ -8,7 +8,7 @@
 
 #include <pcl/features/feature.h>
 
-#include <Ponca/Core/defines.h>
+#include <Ponca/Core>
 
 #include "Eigen/Eigen"
