+++ examples/cpp/pcl/pcl_wrapper.hpp
@@ -8,12 +8,6 @@
 
 #include "pcl_wrapper.h"
 
-#include <Ponca/Core/basket.h>
-#include <Ponca/Core/covariancePlaneFit.h>
-#include <Ponca/Core/curvature.h>
-#include <Ponca/Core/weightFunc.h>
-#include <Ponca/Core/weightKernel.h>
-
 //////////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointInT, typename PointOutT> void
 pcl::GlsCurvature<PointInT, PointOutT>::computeFeature(PointCloudOut &output)
