+++ tests/CMakeLists.txt
@@ -95,7 +95,7 @@ macro(add_multi_test filename)
     add_dependencies(buildtests ${testname})
     foreach(suffix ${suffixes})
       add_executable(${testname}_${suffix} ${filename})
-      target_include_directories(${testname}_${suffix} PRIVATE ${PONCA_ROOT_DIR})
+      target_include_directories(${testname}_${suffix} PRIVATE ${PONCA_src_ROOT})
       ponca_add_target_property(${testname}_${suffix} COMPILE_FLAGS "-DTEST_PART_${suffix}=1")
 
       if(OpenMP_CXX_FOUND)
@@ -110,7 +110,7 @@ macro(add_multi_test filename)
     endforeach(suffix)
   else(suffixes)
     add_executable(${testname} ${filename})
-    target_include_directories(${testname} PRIVATE ${PONCA_ROOT_DIR})
+    target_include_directories(${testname} PRIVATE ${PONCA_src_ROOT})
 
     if(OpenMP_CXX_FOUND)
       target_link_libraries(${testname} PUBLIC OpenMP::OpenMP_CXX)
