+++ tests/src/fit_monge_patch.cpp
@@ -15,11 +15,11 @@
 #include "../common/testing.h"
 #include "../common/testUtils.h"
 
-#include <Ponca/Core/basket.h>
-#include <Ponca/Core/covariancePlaneFit.h>
-#include <Ponca/Core/mongePatch.h>
-#include <Ponca/Core/weightFunc.h>
-#include <Ponca/Core/weightKernel.h>
+#include <Ponca/src/Core/basket.h>
+#include <Ponca/src/Core/covariancePlaneFit.h>
+#include <Ponca/src/Core/mongePatch.h>
+#include <Ponca/src/Core/weightFunc.h>
+#include <Ponca/src/Core/weightKernel.h>
 
 #include <vector>
