+++ tests/src/fit_plane.cpp
@@ -17,11 +17,11 @@
 #include "../common/testing.h"
 #include "../common/testUtils.h"
 
-#include <Ponca/Core/basket.h>
-#include <Ponca/Core/covariancePlaneFit.h>
-#include <Ponca/Core/meanPlaneFit.h>
-#include <Ponca/Core/weightFunc.h>
-#include <Ponca/Core/weightKernel.h>
+#include <Ponca/src/Core/basket.h>
+#include <Ponca/src/Core/covariancePlaneFit.h>
+#include <Ponca/src/Core/meanPlaneFit.h>
+#include <Ponca/src/Core/weightFunc.h>
+#include <Ponca/src/Core/weightKernel.h>
 
 #include <vector>
