+++ tests/src/gls_tau.cpp
@@ -13,12 +13,12 @@
 #include "../common/testing.h"
 #include "../common/testUtils.h"
 
-#include <Ponca/Core/basket.h>
-#include <Ponca/Core/orientedSphereFit.h>
-#include <Ponca/Core/unorientedSphereFit.h>
-#include <Ponca/Core/gls.h>
-#include <Ponca/Core/weightFunc.h>
-#include <Ponca/Core/weightKernel.h>
+#include <Ponca/src/Core/basket.h>
+#include <Ponca/src/Core/orientedSphereFit.h>
+#include <Ponca/src/Core/unorientedSphereFit.h>
+#include <Ponca/src/Core/gls.h>
+#include <Ponca/src/Core/weightFunc.h>
+#include <Ponca/src/Core/weightKernel.h>
 
 #include <vector>
