+++ tests/src/projection.cpp
@@ -16,10 +16,10 @@
 #include "../common/testing.h"
 #include "../common/testUtils.h"
 
-#include <Ponca/Core/basket.h>
-#include <Ponca/Core/orientedSphereFit.h>
-#include <Ponca/Core/weightFunc.h>
-#include <Ponca/Core/weightKernel.h>
+#include <Ponca/src/Core/basket.h>
+#include <Ponca/src/Core/orientedSphereFit.h>
+#include <Ponca/src/Core/weightFunc.h>
+#include <Ponca/src/Core/weightKernel.h>
 
 #include <chrono>
